fixed matching

+++ support_cmp.m
@@ -1,5 +1,5 @@
 function y = support_cmp(one, two)
 
-    s = one-two;
+    s = double(one)-double(two);
     ssd = sum(s(:).^2);
     y = ssd;
